remove unnecessary files
