--- RL_ARC_AGI/mamba_layer.py
@@ -1,6 +1,0 @@
-
-# In actor critic  
-# self.mamba_layers = MambaLayers(d_model=hidden_size,
-#                                 d_state=32,
-#                                 num_layers=3)
-        # shared_features = self.mamba_layers(x)
